sbcl: 32-bit install must get 32-bit bootstrap
Add a resource for downloading the current 32-bit binary
build to use as the bootstrap when installing 32-bit.
Without this sbcl cannot be installed on a 32-bit machine.

With this commit installing on x86_64 machine using the
32-bit option will bootstrap from the 32-bit binary.
Bootstrap to 32-bit from the 64-bit should work, as any boostrap
that can run should produce the same result, but this ensures
at install time that the machine can run the 32-bit version.

Updated the version of the 64-bit bootstrap file
from 1.1.0 to 1.1.8 and added a comment on where to find
documented the list of latest binary versions.

Closes #26850.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>