Determine sqlite include path for python for linux as well.
Python's setup.py extracts include/lib paths from -I, -L sysconfig vars
when the os is darwin, but it even says in the comments it should work
for unixy os as well! The point setup.py tries to build the sqlite3
extension looking for system sqlite3, which either may not exist, or is
not what you want when linking to brew build sqlite3.