Fix bib-tool formula to work with Linuxbrew
The tarball for v. 2.60 of BibTool has a problem where it contains a
duplicate file. This does not play nicely with bsdtar, which is what is
installed on Macs. See https://github.com/Homebrew/homebrew/issues/40559.

The commit that fixed that issue, however, made use of --fast-read (-q),
which is only implemented in bsdtar, not gnutar. In order to make sure
that this formula works with Linuxbrew, we need to define "stage" based
on the flavor of tar that is installed on the machine.