cmake: fix https support in linux by using system curl
It's right to have cmake depend on curl on Linux, but the if/else block
that disables system curl on old OS X versions was actually firing for
Linux as well. The symptom is that all ExternalProject_Add() calls that
used a https url failed to work.

This includes a revision bump because the bottle needs to be rebuilt to
link against the linuxbrew curl (and OpenSSL).